requirements.txt
tweepy==4.14.0
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
loguru==0.7.0
ccxt==4.5.4
aiohttp==3.10.11
fastapi==0.104.1
uvicorn==0.24.0
prometheus-client==0.20.0
google-cloud-logging==3.10.0
google-cloud-secret-manager==2.20.0
